Http Verbs completed

+++ store-management-customers-and-products/src/main/java/com/nata/store_management_customers_and_products/dto/CustomerDtos/CustomerDTO.java
@@ -1,5 +1,7 @@
 package com.nata.store_management_customers_and_products.dto.CustomerDtos;
 
+import java.math.BigDecimal;
+
 public record CustomerDTO (
     String name,    
     String lastname,
@@ -8,7 +10,7 @@ public record CustomerDTO (
     String address,
     String city,
     String state,
-    Double creditlimit
+    BigDecimal creditlimit
 ){}
